requirements.txt
# Web scraping
selenium==4.15.2
praw==7.7.1
beautifulsoup4==4.12.2
newspaper3k==0.2.8
requests==2.31.0

# Data processing
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0

# NLP & ML
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
nltk==3.8.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Crypto data
ccxt==4.1.64
yfinance==0.2.23
